+++ packages/components/src/components/filelist-preview/index.ts
@@ -1,0 +1,5 @@
+import IipFileListPreview from './filelist-preview.vue'
+import { openFileListPreview } from './filelist-preview'
+export * from './types'
+export default IipFileListPreview
+export { IipFileListPreview, openFileListPreview }